refactor: Testes de modelos movidos para a pasta "unit"
